chore: add missing category validations

+++ src/application/usecases/categories/delete_category.ts
@@ -5,6 +5,15 @@ export class DeleteCategory {
 
     constructor(readonly categoryRepository: CategoryRepository) {}
     async execute(categoryId: string, response: Response): Promise<Response> {
+        const categoryResult = await this.categoryRepository.getOne({_id: categoryId})
+
+        if (!categoryResult) {
+            return response.status(404).json({
+                success: true,
+                message: "Category not found",
+            })
+        }
+
         await this.categoryRepository.delete(categoryId);
         return response.status(200).json({
             success: true,
